Tags can have Properties. Much easier to create new tags.

+++ Prototype/Freezing Happiness/BackEnd/Properties/integerproperty.h
@@ -8,7 +8,7 @@ class IntegerProperty : public Property
     Q_OBJECT
 public:
     IntegerProperty(QByteArray *data);
-    IntegerProperty(QString category, QString name, qint64 min, qint64 max, qint64 value = 0);
+    IntegerProperty(QString category, QString name, int min, int max, qint64 value = 0);
     QByteArray toByteArray();
     QString toString() const;
     static QWidget* createWidget(QList<Property *> props, QWidget *parent);
